requirements.txt
Django>=3.0.6
# REST
djangorestframework>=3.11.0
# JWT
djangorestframework-simplejwt>=4.4.0
# PosgreSQL Driver
psycopg2-binary>=2.8.5
# Socket
channels>=2.4.0
# Cheat db link as url
dj-database-url==0.5.0